requirements.txt
appdirs==1.4.4
black==21.5b2
blis==0.7.5
catalogue==2.0.6
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.10
click==7.1.2
cymem==2.0.6
distlib==0.3.1
docopt==0.6.2
fastdiff==0.3.0
filelock==3.0.12
fpdf2==2.2.0
greenlet==1.1.1
idna==3.3
Jinja2==3.0.3
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
mypy-extensions==0.4.3
numpy==1.19.5
opencv-python==4.5.1.48
packaging==21.3
parameterized==0.8.1
pathspec==0.8.1
pathy==0.6.1
Pillow==8.1.0
pipenv==2020.11.15
playwright==1.15.0
preshed==3.0.6
pydantic==1.8.2
pyee==8.2.2
pyparsing==3.0.6
pypdftk==0.5
pysrt==1.1.2
python-dotenv==0.19.0
regex==2020.11.13
requests==2.27.1
rope==0.18.0
six==1.15.0
smart-open==5.2.1
snapshottest==0.6.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
srt==3.5.0
termcolor==1.1.0
thinc==8.0.13
toml==0.10.2
tqdm==4.62.3
typed-ast==1.4.2
typer==0.4.0
typing-extensions==3.7.4.3
urllib3==1.26.8
virtualenv==20.4.4
virtualenv-clone==0.5.4
Wand==0.6.10
wasabi==0.9.0
wasmer==1.1.0
wasmer_compiler_cranelift==1.1.0
websockets==10.0
webvtt-py==0.4.6

# Firebase Integration for Status Updates
firebase-admin==6.2.0
google-cloud-firestore==2.11.1

# AWS S3 Integration for File Handling  
boto3==1.34.131
botocore==1.34.131

# Environment Configuration
python-dotenv==1.0.0

# Local API Server Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
